backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
openai>=1.3.0
anthropic>=0.7.4
celery>=5.3.4
redis>=5.0.1
stripe>=7.0.0
Jinja2>=3.1.2
channels>=4.0.0
channels-redis>=4.0.0
asgiref>=3.5.2
